requirements.txt
# =========================================================
# RAG Enterprise - Dependências
# =========================================================

# LangChain
langchain==0.3.14
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-community==0.3.14
langgraph==0.2.60

# Vector Store
chromadb==0.5.23

# BM25
rank-bm25==0.2.2

# API
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20

# Configuração
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Frontend
streamlit==1.41.1

# Métricas
prometheus-client==0.21.1

# Desenvolvimento
ipykernel==6.29.5
pytest==8.3.4
pytest-asyncio==0.25.2
httpx==0.28.1
